+++ dataforge-data/src/commonMain/kotlin/space/kscience/dataforge/data/DataSink.kt
@@ -11,8 +11,6 @@ import space.kscience.dataforge.names.asName
 import space.kscience.dataforge.names.isEmpty
 import space.kscience.dataforge.names.plus
 
-
-
 /**
  * Asynchronous data sink
  */