+++ dataforge-data/src/commonTest/kotlin/space/kscience/dataforge/data/DataTreeBuilderTest.kt
@@ -13,11 +13,11 @@ internal class DataTreeBuilderTest {
     fun testTreeBuild() = runTest(timeout = 500.milliseconds) {
         val node = DataTree.static<Any> {
             node("primary") {
-                value("a", "a")
-                value("b", "b")
+                putValue("a", "a")
+                putValue("b", "b")
             }
-            value("c.d", "c.d")
-            value("c.f", "c.f")
+            putValue("c.d", "c.d")
+            putValue("c.f", "c.f")
         }
         assertEquals("a", node["primary.a"]?.await())
         assertEquals("b", node["primary.b"]?.await())
@@ -29,16 +29,16 @@ internal class DataTreeBuilderTest {
     @Test
     fun testDataUpdate() = runTest(timeout = 500.milliseconds) {
         val updateData = DataTree.static<Any> {
-            data("a", Data.wrapValue("a"))
-            data("b", Data.wrapValue("b"))
+            put("a", Data.wrapValue("a"))
+            put("b", Data.wrapValue("b"))
         }
 
         val node = DataTree.static<Any> {
             node("primary") {
-                value("a", "a")
-                value("b", "b")
+                putValue("a", "a")
+                putValue("b", "b")
             }
-            value("root", "root")
+            putValue("root", "root")
             node("update", updateData)
         }